Task #230533: fix: Resolved comments and optimize the code
