+++ src/utils/constants/fieldConfig.ts
@@ -1,0 +1,68 @@
+const fieldConfig: any = {
+    username: {
+      KA: {
+        labels: [
+          {
+            code: "FF",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_FIRST_NAME",
+          },
+          {
+            code: "LL",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_LAST_NAME",
+          },
+          { code: "DD", translationKey: "POPUP_DATE" },
+          { code: "MM", translationKey: "POPUP_MONTH" },
+          { code: "YYYY", translationKey: "POPUP_YEAR" },
+        ],
+      },
+      OD: {
+        labels: [
+          {
+            code: "AA",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_FIRST_NAME_OD",
+          },
+          {
+            code: "BB",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_LAST_NAME_OD",
+          },
+          { code: "CC", translationKey: "POPUP_DATE_OD" },
+          { code: "DD", translationKey: "POPUP_MONTH_OD" },
+          { code: "EEEE", translationKey: "POPUP_YEAR_OD" },
+        ],
+      },
+    },
+    password: {
+      KA: {
+        labels: [
+          {
+            code: "FF",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_FIRST_NAME",
+          },
+          {
+            code: "LL",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_LAST_NAME",
+          },
+          { code: "DD", translationKey: "POPUP_DATE" },
+          { code: "MM", translationKey: "POPUP_MONTH" },
+          { code: "YYYY", translationKey: "POPUP_YEAR" },
+        ],
+      },
+      OD: {
+        labels: [
+          {
+            code: "AA",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_FIRST_NAME_OD",
+          },
+          {
+            code: "BB",
+            translationKey: "POPUP_FIRST_TWO_LETTERS_OF_YOUR_LAST_NAME_OD",
+          },
+          { code: "CC", translationKey: "POPUP_DATE_OD" },
+          { code: "DD", translationKey: "POPUP_MONTH_OD" },
+          { code: "EEEE", translationKey: "POPUP_YEAR_OD" },
+        ],
+      },
+    },
+  };
+
+export default fieldConfig;
